# Avoid extra allocation in Encode