Avoid extra allocation to increase performance when encoding to QIPC.

Before
---
```
Starting q process on random port
Listening port is  64211
goos: darwin
goarch: amd64
pkg: github.com/sv/kdbgo
BenchmarkEncodeAll-8   	   30000	     48148 ns/op
PASS
Q stdout output: "show `exiting_process;\n"

Q stdout output: `exiting_process

Q stdout output: "exit 0\n"

Q stdout output:
Q stdout error: EOF
Q stderr output:
Q stderr error: EOF
ok  	github.com/sv/kdbgo	2.013s
```

After
---
```
Starting q process on random port
Listening port is  52052
goos: darwin
goarch: amd64
pkg: github.com/sv/kdbgo
BenchmarkEncodeAll-8   	  100000	     20985 ns/op
PASS
Q stdout output: "show `exiting_process;\n"

Q stdout output: `exiting_process
"exit 0\n"

Q stdout output:
Q stdout error: EOF
Q stderr output:
Q stderr error: EOF
ok  	github.com/sv/kdbgo	2.403s
```