@sv 2019-03-01
Could you update description with new test timing if it has changed?

On Fri, 1 Mar 2019 at 16:30 Nima Yahyazadeh <notifications@github.com>
wrote:

> @sv <https://github.com/sv> Could you re-review
>
> —
> You are receiving this because you were mentioned.
>
>
> Reply to this email directly, view it on GitHub
> <https://github.com/sv/kdbgo/pull/20#issuecomment-468724784>, or mute the
> thread
> <https://github.com/notifications/unsubscribe-auth/AAA3Brd8GVHX9OCOMJAIBkBXxvFXrwYmks5vSVWngaJpZM4bYLFF>
> .
>