@yznima 2019-03-01
> Could you update description with new test timing if it has changed?

@sv Done